requirements.txt
autopep8>=1.6.0
click>=8.0.3
coverage>=5.5
email_validator>=1.1.3
Faker>=11.3.0
Flask>=2.0
flask-cors>=3.0.10
flask-FormRequest>=0.0.23
Flask-Migrate>=3.1.0
Flask-Seeder>=1.2.0
Flask-SQLAlchemy>=2.5.1
Flask-WTF>=0.15.1
gunicorn>=20.1
psycopg2>=2.9.1
pyjwt>=2.3.0
pytest>=6.2
python-dotenv>=0.19.0
